Moved tests for shifted functions from test_transform to test_convert, added angular type checking to various functions in geodepy.convert
